Run grip on 0.0.0.0 without browser open on CTRL-F5
Noraml F5 open grip on the current markdown file, binding to localhost
and also auto opens the browser.
For running from a container, this does not make sense since you do not
normally want to run a browser in the container, and you also mostyl
want to bind to 0.0.0.0 so you can access the grip server from the host.
This is what the CTRL-F5 vim binding now does for the current markdown
file.